+++ src/components/Indices/Index.vue
@@ -13,7 +13,7 @@
           <h2 class="subheading">{{indexName}}</h2>
           <data-loader method="indicesGet" :methodParams="{index: indexName}" ref="dataLoader">
             <template slot-scope="data">
-              <print-pretty-or-raw :document="data.body"></print-pretty-or-raw>
+              <print-pretty :document="data.body"></print-pretty>
             </template>
           </data-loader>
         </v-card-text>
@@ -24,7 +24,7 @@
 
 <script>
   import BackButton from '@/components/shared/BackButton'
-  import PrintPrettyOrRaw from '@/components/shared/PrintPrettyOrRaw'
+  import PrintPretty from '@/components/shared/PrintPretty'
 
   export default {
     name: 'Index',
@@ -34,7 +34,7 @@
       }
     },
     components: {
-      PrintPrettyOrRaw,
+      PrintPretty,
       BackButton
     }
   }