+++ src/components/Indices/IndicesTable.vue
@@ -3,10 +3,11 @@
     <v-card-text>
       <v-flex right d-inline-flex>
         <v-text-field append-icon="search"
-                      v-on:keyup.esc="indicesFilter = ''"
+                      @keyup.esc="indicesFilter = ''"
                       label="Filter via 'column:query'"
                       name="filter"
                       id="filter"
+                      autofocus
                       v-model="indicesFilter"></v-text-field>
       </v-flex>
     </v-card-text>
@@ -73,7 +74,10 @@
     },
     props: {
       indices: {
-        default: () => []
+        default: () => {
+          return []
+        },
+        type: Array
       },
       loading: {
         default: false