+++ src/components/Setup/ConnectBase.vue
@@ -66,6 +66,11 @@
             this.connectError = true
             this.showErrorSnackbar({text: 'Error: could not connect.'})
           })
+      },
+      connectIfTestSuccessfull () {
+        if (this.testSuccess) {
+          this.connect()
+        }
       }
     },
     mixins: [ConnectWithServer]